requirements.txt
ollama>=0.1.7
httpx>=0.27.0
aiohttp>=3.9.5
numpy>=1.26.0
faiss-cpu>=1.7.4
gradio>=5.0.0

